Added Camera functionality B)
