+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/adampkg/auto/RedLeftBlueRight.java
@@ -6,7 +6,7 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.Pose2D;
-import org.firstinspires.ftc.teamcode.adampkg.RobotBase;
+import org.firstinspires.ftc.teamcode.adampkg.teleop.RobotBase;
 
 import java.util.Locale;
 @Autonomous(name="RedLeft_BlueLeft_AQ", group="VECTORAUTO")
