+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/adampkg/teleop/RobotBase.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.adampkg;
+package org.firstinspires.ftc.teamcode.adampkg.teleop;
 
 //import com.qualcomm.robotcore.eventloop.opmode.OpModeInternal;
 //import com.qualcomm.robotcore.eventloop.opmode.OpModeInternal;
