+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/adampkg/teleop/StatesTeleop.java
@@ -1,20 +1,24 @@
 
-package org.firstinspires.ftc.teamcode.adampkg;
+package org.firstinspires.ftc.teamcode.adampkg.teleop;
 
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
-
 //@Disabled
-@TeleOp(group = "Qureshi", name = "ManualTeleop")
-public class MechAnimAQ extends LinearOpMode {
-
+@TeleOp(group = "Qureshi", name = "StatesTeleop")
+public class StatesTeleop extends LinearOpMode {
+    int MAX_HEIGHT = 6130;
+    int SCORING_HEIGHT = 2575;
+    int DOWN_SCORING_HEIGHT = 1850;
     final RobotBase robotBase = new RobotBase();
 
     boolean slowMode = false;
+    boolean slowModeDebounce = false;
 
-    double debounceTime = time;
+    boolean liftTicksDebounce = false;
+    int liftTicks;
+
+    boolean dpadDebounce = false;
 
     @Override
     public void runOpMode() throws InterruptedException {
@@ -23,38 +27,120 @@ public class MechAnimAQ extends LinearOpMode {
         }
         waitForStart();
         while (opModeIsActive()) {
+            telemetry.addData("Touch Sensor:", robotBase.touchSensor.getValue());
+            tickReset();
             servoTelemetry();
             teleOpControls();
+            telemetry.update();
         }
 
     }
+    public void tickReset()
+    {
+        if(robotBase.touchSensor.isPressed())
+        {
+            initMotors();
+            telemetry.addData("Is pressed", robotBase.touchSensor.isPressed());
+            liftTicks = 0;
+        }
+    }
+    public void initMotors()
+    {
+        robotBase.liftLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        robotBase.liftRight.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+    }
+
+    public void runLiftsToPos(int position)
+    {
+      //  robotBase.initServos(hardwareMap);
+        robotBase.liftLeft.setTargetPosition(position);
+        robotBase.liftRight.setTargetPosition(position);
+        robotBase.liftLeft.setPower(1);
+        robotBase.liftRight.setPower(1);
+        robotBase.liftLeft.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        robotBase.liftRight.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+    }
+    public void  motorTelemetry(){
+        telemetry.addData("leftLift","Encoder: %2d, Power: %2f", robotBase.liftLeft.getCurrentPosition(), robotBase.liftLeft.getPower());
+        telemetry.addData("leftRight","Encoder: %2d, Power: %2f", robotBase.liftRight.getCurrentPosition(), robotBase.liftRight.getPower());
+        telemetry.update();
+    }
+    public void waitLifts(int holdTime)
+    {
+        while(opModeIsActive() && robotBase.liftRight.isBusy() && robotBase.liftLeft.isBusy())
+        {
+            motorTelemetry();
+        }
+        sleep(holdTime);
+    }
     public void Lift()
     {
+        telemetry.addData("Left Tick number:", robotBase.liftLeft.getCurrentPosition());
+        telemetry.addData("Right Tick number:", robotBase.liftRight.getCurrentPosition());
+
+        if (gamepad1.dpad_left)
+        {
+            if(!dpadDebounce) {
+                runLiftsToPos(SCORING_HEIGHT);
+                dpadDebounce = true;
+            }
+
+        }
+        else if (gamepad1.dpad_up)
+        {
+            if(!dpadDebounce) {
+                runLiftsToPos(MAX_HEIGHT);
+                dpadDebounce = true;
+            }
+        }
+        else if (gamepad1.dpad_down) {
+            if(!dpadDebounce) {
+                runLiftsToPos(0);
+                dpadDebounce = true;
+            }
+        }
+        else if (gamepad1.dpad_right){
+            if(!dpadDebounce) {
+                runLiftsToPos(DOWN_SCORING_HEIGHT);
+                dpadDebounce = true;
+            }
+        } else {
+            dpadDebounce = false;
+        }
+
         if(gamepad1.y)
         {
-            robotBase.liftLeft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            robotBase.liftRight.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+            liftTicksDebounce = false;
+            robotBase.liftLeft.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+            robotBase.liftRight.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
             robotBase.liftDirection = 1;
         }
         else if(gamepad1.a)
         {
-            robotBase.liftLeft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            robotBase.liftRight.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+            liftTicksDebounce = false;
+            robotBase.liftLeft.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+            robotBase.liftRight.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
             robotBase.liftDirection = -1;
         }
         else {
-            robotBase.liftLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-            robotBase.liftRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-
-            robotBase.liftDirection = 0;
+            if(!liftTicksDebounce){
+                liftTicks = robotBase.liftLeft.getCurrentPosition();
+                liftTicksDebounce = true;
+                runLiftsToPos(liftTicks);
+            }
+            robotBase.liftDirection = 0.5;
+          //  runLiftsToPos(liftTicks);
         }
         robotBase.liftRight.setPower(robotBase.liftDirection);
         robotBase.liftLeft.setPower(robotBase.liftDirection);
+        telemetry.addData("liftTicks", liftTicks);
+
 
         //return liftDirection;
     }
     public void initHardware() {
         initDrive();
+        initMotors();
         robotBase.initServos(hardwareMap);
     }
 
@@ -103,11 +189,13 @@ public class MechAnimAQ extends LinearOpMode {
     public void slowModeSwitcher()
     {
         if(gamepad1.left_stick_button){
-            if( time >= debounceTime + 10000) {
+            if(!slowModeDebounce){
                 slowMode = !slowMode;
-                debounceTime = time;
+                slowModeDebounce = true;
             }
-
+        }else{
+            slowModeDebounce = false;
+         //   telemetry.addData("stick", "stick up");
         }
     }
     public void driveTrain()
