+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/adampkg/teleop/UtilityCameraFrameCapture.java
@@ -1,0 +1,111 @@
+package org.firstinspires.ftc.teamcode.adampkg.teleop;
+
+
+import android.util.Size;
+
+
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+
+
+import org.firstinspires.ftc.robotcore.external.hardware.camera.BuiltinCameraDirection;
+import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
+import org.firstinspires.ftc.vision.VisionPortal;
+
+
+import java.util.Locale;
+
+
+/*
+ * This OpMode helps calibrate a webcam or RC phone camera, useful for AprilTag pose estimation
+ * with the FTC VisionPortal.   It captures a camera frame (image) and stores it on the Robot Controller
+ * (Control Hub or RC phone), with each press of the gamepad button X (or Square).
+ * Full calibration instructions are here:
+ *
+ *  https://ftc-docs.firstinspires.org/camera-calibration
+ *
+ * In Android Studio, copy this class into your "teamcode" folder with a new name.
+ * Remove or comment out the @Disabled line to add this OpMode to the Driver Station OpMode list.
+ *
+ * In OnBot Java, use "Add File" to add this OpMode from the list of Samples.
+ */
+
+
+@TeleOp(name = "Utility: Camera Frame Capture", group = "Utility")
+public class UtilityCameraFrameCapture extends LinearOpMode
+{
+    /*
+     * EDIT THESE PARAMETERS AS NEEDED
+     */
+    final boolean USING_WEBCAM = true;
+    final BuiltinCameraDirection INTERNAL_CAM_DIR = BuiltinCameraDirection.BACK;
+    final int RESOLUTION_WIDTH = 640;
+    final int RESOLUTION_HEIGHT = 480;
+
+
+    // Internal state
+    boolean lastX;
+    int frameCount;
+    long capReqTime;
+
+
+    @Override
+    public void runOpMode()
+    {
+        VisionPortal portal;
+
+
+        if (USING_WEBCAM)
+        {
+            portal = new VisionPortal.Builder()
+                    .setCamera(hardwareMap.get(WebcamName.class, "kyleCamLeft"))
+                    .setCameraResolution(new Size(RESOLUTION_WIDTH, RESOLUTION_HEIGHT))
+                    .build();
+        }
+        else
+        {
+            portal = new VisionPortal.Builder()
+                    .setCamera(INTERNAL_CAM_DIR)
+                    .setCameraResolution(new Size(RESOLUTION_WIDTH, RESOLUTION_HEIGHT))
+                    .build();
+        }
+
+
+        while (!isStopRequested())
+        {
+            boolean x = gamepad1.x;
+
+
+            if (x && !lastX)
+            {
+                portal.saveNextFrameRaw(String.format(Locale.US, "CameraFrameCapture-%06d", frameCount++));
+                capReqTime = System.currentTimeMillis();
+            }
+
+
+            lastX = x;
+
+
+            telemetry.addLine("######## Camera Capture Utility ########");
+            telemetry.addLine(String.format(Locale.US, " > Resolution: %dx%d", RESOLUTION_WIDTH, RESOLUTION_HEIGHT));
+            telemetry.addLine(" > Press X (or Square) to capture a frame");
+            telemetry.addData(" > Camera Status", portal.getCameraState());
+
+
+            if (capReqTime != 0)
+            {
+                telemetry.addLine("\nCaptured Frame!");
+            }
+
+
+            if (capReqTime != 0 && System.currentTimeMillis() - capReqTime > 1000)
+            {
+                capReqTime = 0;
+            }
+
+
+            telemetry.update();
+        }
+    }
+}
+
