--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/edmondpkg/readme.md
@@ -1,6 +1,0 @@
-Hello! You are in edmondpkg. This is the package that I, Edmond Zalkin have created for testing code.
-
-File contents:
-    BELTDRIVETEST -- This file is the basic teleop code for a belt driven robot with joystick controls
-                     on one controller.
-    More files will be included in the future.
